+++ libzutil/udp.h
@@ -14,8 +14,8 @@ extern "C" {
 
 intptr_t udp_open(char *src_addr, char *addr, unsigned short port, char *mode);
 int udp_close(intptr_t id);
-size_t udp_read(intptr_t id, void *buf);
-size_t udp_write(intptr_t id, const void *buf, int len);
+ssize_t udp_read(intptr_t id, void *buf);
+ssize_t udp_write(intptr_t id, const void *buf, size_t len);
 
 #ifdef __cplusplus
 }
